Move new Pane renaming to folder
