+++ static/panes/pane.ts
@@ -32,7 +32,7 @@ import { FontScale } from '../widgets/fontscale';
 import { SiteSettings } from '../settings';
 import * as utils from '../utils';
 
-import { PaneRenaming } from '../pane-renaming';
+import { PaneRenaming } from '../widgets/pane-renaming';
 
 /**
  * Basic container for a tool pane in Compiler Explorer
